cylc monitor: migrate into cylc cli
